// frontend/generated/models/ImageDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 8lind8box Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AccountDto } from './AccountDto';
import {
    AccountDtoFromJSON,
    AccountDtoFromJSONTyped,
    AccountDtoToJSON,
    AccountDtoToJSONTyped,
} from './AccountDto';

/**
 * 
 * @export
 * @interface ImageDto
 */
export interface ImageDto {
    /**
     * 
     * @type {number}
     * @memberof ImageDto
     */
    imageId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof ImageDto
     */
    uploader?: AccountDto;
    /**
     * 
     * @type {number}
     * @memberof ImageDto
     */
    blindBoxId?: number;
    /**
     * 
     * @type {number}
     * @memberof ImageDto
     */
    toyId?: number;
    /**
     * 
     * @type {number}
     * @memberof ImageDto
     */
    skuId?: number;
    /**
     * 
     * @type {string}
     * @memberof ImageDto
     */
    imageUrl?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ImageDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ImageDto
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the ImageDto interface.
 */
export function instanceOfImageDto(value: object): value is ImageDto {
    return true;
}

export function ImageDtoFromJSON(json: any): ImageDto {
    return ImageDtoFromJSONTyped(json, false);
}

export function ImageDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ImageDto {
    if (json == null) {
        return json;
    }
    return {
        
        'imageId': json['imageId'] == null ? undefined : json['imageId'],
        'uploader': json['uploader'] == null ? undefined : AccountDtoFromJSON(json['uploader']),
        'blindBoxId': json['blindBoxId'] == null ? undefined : json['blindBoxId'],
        'toyId': json['toyId'] == null ? undefined : json['toyId'],
        'skuId': json['skuId'] == null ? undefined : json['skuId'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function ImageDtoToJSON(json: any): ImageDto {
    return ImageDtoToJSONTyped(json, false);
}

export function ImageDtoToJSONTyped(value?: ImageDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'imageId': value['imageId'],
        'uploader': AccountDtoToJSON(value['uploader']),
        'blindBoxId': value['blindBoxId'],
        'toyId': value['toyId'],
        'skuId': value['skuId'],
        'imageUrl': value['imageUrl'],
        'isVisible': value['isVisible'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
    };
}